implement pruning for plugins
